prog: implement compressed_image generation
We have 21 filesystems that don't have seeds and it's hard to generate them.
As the result we stopped testing them entirely.
Implement generation of compressed_image that produces just an empty image.
This will re-enable testing of all filesystems.
We had a similar issue with syz_read_part_table. While we added seeds for it,
it looks safer to allow generation of compressed_image in general to prevent
such issues in future.