prog: better mutation of empty images
If we start with an empty image, we need to be able to mutate all bytes,
not just the "interesting" ones.
Consider we start with an empty image and flip a single byte in it.
The next time we will see it's not trivial and create a single
64-byte interesting segment. If we only mutate "interesting" data,
we will never mutate anything else in this image.

Also adjust number of mutations per image.
We do one per 128K of interesting data. But we have no single seed
that has so much interesting data. Here is the amount of interesting data in seeds:
https://gist.github.com/dvyukov/566b20364610d80e5d0534524546c3f0
Reduce this cuhnk size to 4K.