pkg/fuzzer: simplify choice table updates
It's better to rebuild choice table when we add new inputs
to the corpus rather than every time we query the table.
The condition for regeneration depends only on the corpus size.
Querying happens much more frequently, so we can remove
mutex/check overhead from that path.

This also removes the need in the separate gorotuine,
the notification channel and the tricky lossy prototcol.
I think currently we can also regenerate it twice in a row.
Regenerating in the triage job is much simpler.