tools/syz-declextract: extract file_operations descriptions
Emit descriptions for special files in /dev, /sys, and /proc.

First, the clang tool locates file_operations variables and arrays
and dumps open/read/write/mmap/ioctl callbacks for each.
It also tries to extract set of ioctl commands and argument types
for them in a simple best-effort way (for now). It just locates switch
in the ioctl callback and extracts each case as a command.

Then, dynamic probing opens all files in dev/sys/proc,
and dumps symbolized coverage with function names for each file.

Then, the Go tool parses both sources of the information,
maps file names in dev/sys/proc to the file_operations callbacks
using function names in coverage, and emits open/read/write/mmap/ioctl
descriptions.

The clang tool dumps information in the new JSON format.
It's designed to be extensible, the idea is that in the future
we switch all types of information to the JSON format
(all output is a sequence of JSON records of different types).

The descriptions are extracted on the kernel commit 228a1157fb9f
with syzbot smack config. They are not perfect yet, but it's
a good start.